+++ intl/msg_hash_us.c
@@ -109,6 +109,10 @@ int msg_hash_get_help_us_enum(enum msg_hash_enums msg, char *s, size_t len)
              snprintf(s, len,
                    "Toggles frames per second counter.");
              break;
+          case RARCH_STATISTICS_TOGGLE:
+             snprintf(s, len,
+                   "Toggles display of technical statistics.");
+             break;
           case RARCH_SEND_DEBUG_INFO:
              snprintf(s, len,
                    "Sends diagnostic info about your device and RetroArch configuration to our servers for analysis.");